pyproject.toml
[project]
name = "airflow-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow",
    "graphviz>=0.20.3",
    "pandas>=2.2.3",
]
